fix: refactoring GamePage.js, add DrawingCanvas.js
